Add files via upload
-Mesh Assets added
-rotate camera by moving mouse to edge of screen
-left click to move forward
-the lights appearing at you feet is a tracer of everything you walk on
-the "cheat" light can be enable/disabled for debug (game not playable yet when disabled)

P.S. I had to setup without Google Cardboard for now because of camera control issue. We can easily port it later. I just want to get this first build working as a tech demo for the presentation.
P.P.S. I will demo this software to you guys in more detail when we meet on Wednesday.